+++ rotaryDecoder.h
@@ -2,7 +2,7 @@
 //
 //    FILE: rotaryDecoder.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.1
+// VERSION: 0.3.0
 //    DATE: 2021-05-08
 // PURPOSE: rotary decoder library for Arduino
 //     URL: https://github.com/RobTillaart/rotaryDecoder
@@ -37,11 +37,17 @@ public:
   int32_t getValue(uint8_t re);
   void    setValue(uint8_t re, int32_t value = 0);
 
+
+  //  READ - WRITE interface
+  uint8_t read1(uint8_t pin);
+  bool    write1(uint8_t pin, uint8_t value);
+
+
   //  DEBUG
   uint8_t getLastPosition(uint8_t re);
 
 
-private:
+protected:
   uint8_t   _count = 0;
   uint8_t   _lastValue = 0;
   uint8_t   _lastPos[4] = { 0, 0, 0, 0 };